# be/docs/swagger.yaml
definitions:
...
  codon-backend_internal_models.Course:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  codon-backend_internal_models.KYCStatus:
    enum:
    - not_required
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - KYCNotRequired
    - KYCPending
    - KYCApproved
    - KYCRejected
...
  codon-backend_internal_models.Session:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      device_info:
        type: string
      expires_at:
        type: string
      id:
        description: also the JWT jti
        type: string
      last_used_at:
        type: string
      revoked_at:
        type: string
      user_id:
        type: string
    type: object
...
  codon-backend_internal_models.Subscription:
    properties:
      auto_renew:
        type: boolean
      course:
        $ref: '#/definitions/codon-backend_internal_models.Course'
      course_id:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      plan:
        $ref: '#/definitions/codon-backend_internal_models.SubscriptionPlan'
      plan_id:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/codon-backend_internal_models.SubscriptionStatus'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  codon-backend_internal_models.SubscriptionPlan:
    properties:
      benefits:
        items:
          type: string
        type: array
      course:
        $ref: '#/definitions/codon-backend_internal_models.Course'
      course_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      duration_days:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price_paise:
        type: integer
      updated_at:
        type: string
    type: object
  codon-backend_internal_models.SubscriptionStatus:
    enum:
    - pending_payment
    - active
    - expired
    - cancelled
    type: string
    x-enum-varnames:
    - SubPendingPayment
    - SubActive
    - SubExpired
    - SubCancelled
...
  codon-backend_internal_models.User:
    properties:
      can_manage_all_content:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      kyc_status:
        $ref: '#/definitions/codon-backend_internal_models.KYCStatus'
      last_login_at:
        type: string
      name:
        type: string
      phone_number:
        type: string
      profile_photo_key:
        type: string
      role:
        $ref: '#/definitions/codon-backend_internal_models.UserRole'
      selected_course:
        $ref: '#/definitions/codon-backend_internal_models.Course'
      selected_course_id:
        type: string
      updated_at:
        type: string
    type: object
...
  codon-backend_internal_models.UserRole:
    enum:
    - student
    - teacher
    - admin
    type: string
    x-enum-varnames:
    - RoleStudent
    - RoleTeacher
    - RoleAdmin
...
  internal_handlers.adminUserSessionsResponse:
    properties:
      active_count:
        type: integer
      sessions:
        items:
          $ref: '#/definitions/codon-backend_internal_models.Session'
        type: array
    type: object
  internal_handlers.adminUserSubscriptionResponse:
    properties:
      subscription:
        $ref: '#/definitions/codon-backend_internal_models.Subscription'
    type: object
  internal_handlers.analyticsOverviewResponse:
    properties:
      active_subscriptions:
        type: integer
      active_subscriptions_delta:
        type: string
      active_subscriptions_positive:
        type: boolean
      new_users:
        type: integer
      new_users_delta:
        type: string
      new_users_positive:
        type: boolean
      pending_reviews:
        type: integer
      published_brain_hacks:
        type: integer
      published_content:
        type: integer
      published_tests:
        type: integer
      revenue:
        type: integer
      revenue_bars:
        items:
          type: integer
        type: array
      revenue_delta:
        type: string
      revenue_positive:
        type: boolean
      student_bars:
        items:
          type: integer
        type: array
      teacher_bars:
        items:
          type: integer
        type: array
      tests_attempted:
        type: integer
      tests_attempted_delta:
        type: string
      tests_attempted_positive:
        type: boolean
    type: object
...
  internal_handlers.dashboardSummaryResponse:
    properties:
      active_subscriptions:
        example: 830
        type: integer
      pending_content_reviews:
        example: 3
        type: integer
      pending_kyc:
        example: 12
        type: integer
      pending_test_reviews:
        example: 5
        type: integer
      total_users:
        example: 1420
        type: integer
    type: object
...
  internal_handlers.errorResponse:
    properties:
      code:
        example: pool_too_small
        type: string
      details: {}
      error:
        example: descriptive error message
        type: string
    type: object
...
  internal_handlers.listUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/codon-backend_internal_models.User'
        type: array
    type: object
...
  internal_handlers.messageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  internal_handlers.updateRoleRequest:
    properties:
      role:
        enum:
        - student
        - teacher
        - admin
        example: teacher
        type: string
    type: object
...
  internal_handlers.updateTeacherPermissionsRequest:
    properties:
      can_manage_all_content:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /api/v1/admin/analytics:
    get:
      description: Returns analytics data for charts and stats, with computed deltas
        based on date ranges
      parameters:
      - default: 7D
        description: Time range (7D, 30D, All Time)
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.analyticsOverviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get admin analytics overview
      tags:
      - Admin
...
  /api/v1/admin/dashboard/summary:
    get:
      description: 'Returns live counts for the admin dashboard: total users, active
        subscriptions, pending KYC records, and pending content review items.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.dashboardSummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get admin dashboard summary
      tags:
      - Admin
...
  /api/v1/admin/users:
    get:
      description: Returns all users filterable by role and searchable by phone number
        or name.
      parameters:
      - description: 'Filter by role: student | teacher | admin'
        in: query
        name: role
        type: string
      - description: Search by phone number or name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.listUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List all users (Admin)
      tags:
      - Admin
  /api/v1/admin/users/{id}:
    get:
      description: Returns a single user's profile including their selected course.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codon-backend_internal_models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a user by ID (Admin)
      tags:
      - Admin
...
  /api/v1/admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: 'Changes a user''s role. Role is always admin-managed; users cannot
        self-assign roles. Valid roles: student, teacher, admin.'
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a user's role (Admin)
      tags:
      - Admin
  /api/v1/admin/users/{id}/sessions:
    get:
      description: Returns the list of active (non-revoked, non-expired) device sessions
        for a user.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.adminUserSessionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a user's active sessions (Admin)
      tags:
      - Admin
  /api/v1/admin/users/{id}/subscription:
    get:
      description: Returns the active subscription for the specified user, or null
        if none.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.adminUserSubscriptionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a user's active subscription (Admin)
      tags:
      - Admin
  /api/v1/admin/users/{id}/teacher-permissions:
    patch:
      consumes:
      - application/json
      description: Grants or revokes the can_manage_all_content flag for a teacher.
        When true, the teacher can edit and moderate all content platform-wide regardless
        of authorship.
      parameters:
      - description: User UUID (must be a teacher)
        in: path
        name: id
        required: true
        type: string
      - description: Permission flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateTeacherPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update teacher content permissions (Admin)
      tags:
      - Admin
...
swagger: "2.0"